+++ example/simple-api/tsp-output/Helpers/MergePatchValue.g.cs
@@ -11,6 +11,7 @@
 using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+
 namespace MassiveScale.Simple.Api.Example.Helpers
 {
     /// <summary>