+++ example/simple-api/tsp-output/Services/IWidgetsService.g.cs
@@ -11,7 +11,8 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-namespace MassiveScale.Simple.Api.Example.Services
+
+namespace MassiveScale.Simple.Api.Example
 {
     public interface IWidgetsService
     {